requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
constantly==15.1.0
cryptography==41.0.2
cymem==2.0.7
daphne==4.0.0
Django==4.2.2
django-cors-headers==3.13.0
django-uuidfield==0.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.9
numpy==1.24.4
packaging==23.1
pathy==0.10.2
Pillow==9.4.0
preshed==3.0.8
psycopg2==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.6.0
pyOpenSSL==23.2.0
pytz==2022.7.1
reportlab==4.0.4
requests==2.28.2
service-identity==23.1.0
six==1.16.0
smart-open==6.3.0
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
thinc==8.1.10
tqdm==4.65.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typer==0.9.0
typing_extensions==4.7.0
tzdata==2022.7
urllib3==1.26.14
wasabi==1.1.2
zope.interface==6.0
